Move get_horizontal_elem_size_3d to utility3d
