+++ thetis/utility.py
@@ -526,18 +526,6 @@ def get_horizontal_elem_size_2d(sol2d):
     solve(a == l, sol2d)
 
 
-def get_horizontal_elem_size_3d(sol2d, sol3d):
-    """
-    Computes horizontal element size from the 2D mesh, then copies it on a 3D
-    field
-
-    :arg sol2d: 2D :class:`Function` for the element size field
-    :arg sol3d: 3D :class:`Function` for the element size field
-    """
-    get_horizontal_elem_size_2d(sol2d)
-    ExpandFunctionTo3d(sol2d, sol3d).solve()
-
-
 def get_facet_areas(mesh):
     """
     Compute area of each facet of `mesh`. The facet areas are stored as a HDiv trace field.
